Expose expressions for C API (#7190)
Expose a subset of expressions to test filtering in Scan API:
root, column, get_item, select, or, not, is_null, and binary operators,
including shortcuts like eq.
Add vx_array_apply to test C-side expression application.

This patch does not include C++ FFI tests because we can't create
vx_arrays' yet, #7148.
The next steps are exposing literal expressions and dynamic expressions.

+++ vortex-ffi/src/file.rs
@@ -185,8 +185,10 @@ pub unsafe extern "C-unwind" fn vx_file_open_reader(
             .parse()
             .map_err(|e| vortex_err!("Failed to parse URI '{}': {}", uri_str, e))?;
 
-        let prop_keys = unsafe { to_string_vec(options.property_keys, options.property_len) };
-        let prop_vals = unsafe { to_string_vec(options.property_vals, options.property_len) };
+        let prop_keys =
+            unsafe { to_string_vec(options.property_keys, options.property_len as usize) };
+        let prop_vals =
+            unsafe { to_string_vec(options.property_vals, options.property_len as usize) };
 
         let object_store = make_object_store(&uri, &prop_keys, &prop_vals)?;
